+++ event-store-management/event-store-management-core/src/main/java/uk/gov/justice/services/eventstore/management/catchup/process/SubscriptionCatchupDetailsMapper.java
@@ -12,12 +12,16 @@ public class SubscriptionCatchupDetailsMapper {
     @Inject
     private PriorityComparatorProvider priorityComparatorProvider;
 
+    @Inject
+    private EventSourceNameFilter eventSourceNameFilter;
+
     public Stream<SubscriptionCatchupDetails> toSubscriptionCatchupDetails(final SubscriptionsDescriptor subscriptionsDescriptor) {
 
         final String componentName = subscriptionsDescriptor.getServiceComponent();
 
         return subscriptionsDescriptor.getSubscriptions()
                 .stream()
+                .filter(subscription -> eventSourceNameFilter.shouldRunCatchupAgainstEventSource(subscription))
                 .sorted(priorityComparatorProvider.getSubscriptionComparator())
                 .map(subscription ->
                         new SubscriptionCatchupDetails(